+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Added
 - `maelstrom-go-test` now edits out some lines from test failures that come from the `go test` test
   fixture code.
+- `maelstrom-go-test` now shows tests which call `t.Skip` as status `IGNORED`
 
 ## [0.11.0] - 2024-07-30